+++ NServiceBus.FileBasedRouting/FileBasedRoutingFeature.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Xml;
+using System.Xml.Linq;
 using NServiceBus.Features;
 using NServiceBus.Routing;
 
@@ -21,7 +23,13 @@ namespace NServiceBus.FileBasedRouting
         protected override void Setup(FeatureConfigurationContext context)
         {
             var unicastRoutingTable = context.Settings.Get<UnicastRoutingTable>();
-            var routingFile = new XmlRoutingFile(()=> File.OpenRead(context.Settings.Get<string>(RoutingFilePathKey)));
+            XDocument document;
+            using (var fileStream = File.OpenRead(context.Settings.Get<string>(RoutingFilePathKey)))
+            {
+                document = XDocument.Load(fileStream);
+            }
+
+            var routingFile = new XmlRoutingFileParser(document);
 
             // ensure the routing file is valid and the routing table is populated before running FeatureStartupTasks
             UpdateRoutingTable(routingFile, unicastRoutingTable);
@@ -29,9 +37,9 @@ namespace NServiceBus.FileBasedRouting
             context.RegisterStartupTask(new UpdateRoutingTask(routingFile, unicastRoutingTable));
         }
 
-        private static void UpdateRoutingTable(XmlRoutingFile routingFile, UnicastRoutingTable unicastRoutingTable)
+        private static void UpdateRoutingTable(XmlRoutingFileParser routingFileParser, UnicastRoutingTable unicastRoutingTable)
         {
-            var endpoints = routingFile.Read();
+            var endpoints = routingFileParser.Read();
 
             var commandRoutes = new List<RouteTableEntry>();
             foreach (var endpoint in endpoints)
@@ -48,19 +56,19 @@ namespace NServiceBus.FileBasedRouting
 
         class UpdateRoutingTask : FeatureStartupTask, IDisposable
         {
-            private readonly XmlRoutingFile routingFile;
+            private readonly XmlRoutingFileParser routingFileParser;
             private readonly UnicastRoutingTable unicastRoutingTable;
             private Timer updateTimer;
 
-            public UpdateRoutingTask(XmlRoutingFile routingFile, UnicastRoutingTable unicastRoutingTable)
+            public UpdateRoutingTask(XmlRoutingFileParser routingFileParser, UnicastRoutingTable unicastRoutingTable)
             {
-                this.routingFile = routingFile;
+                this.routingFileParser = routingFileParser;
                 this.unicastRoutingTable = unicastRoutingTable;
             }
 
             protected override Task OnStart(IMessageSession session)
             {
-                updateTimer = new Timer(state => UpdateRoutingTable(routingFile, unicastRoutingTable), null, TimeSpan.FromSeconds(30), TimeSpan.FromSeconds(30));
+                updateTimer = new Timer(state => UpdateRoutingTable(routingFileParser, unicastRoutingTable), null, TimeSpan.FromSeconds(30), TimeSpan.FromSeconds(30));
 
                 return Task.CompletedTask;
             }
