+++ NServiceBus.FileBasedRouting/NServiceBus.FileBasedRouting.csproj
@@ -47,7 +47,7 @@
     <Compile Include="EndpointRoutingConfiguration.cs" />
     <Compile Include="FileBasedRoutingFeature.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="XmlRoutingFile.cs" />
+    <Compile Include="XmlRoutingFileParser.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="routing.xsd">
